livekit-scribe/canvas-plugin/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "livekit_scribe"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "canvas>=0.112.0",
    "livekit-api>=1.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["livekit_scribe"]
